chore: move steps to a folder
