+++ tests/cypress/support/step_definitions/jupyterlite/I Run All Cells.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I Run All Cells", () => {
     const { jupyterLiteSession } = new MaterialDesignerPage().designerWidget;
