+++ tests/cypress/support/step_definitions/jupyterlite/I click on link.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I click on {string} link", (link: string) => {
     new MaterialDesignerPage().designerWidget.jupyterLiteSession.clickOnLink(link);
