+++ tests/cypress/support/step_definitions/jupyterlite/I open JupyterLite Transformation dialog.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I open JupyterLite Transformation dialog", () => {
     new MaterialDesignerPage().designerWidget.headerMenu.selectMenuItemByNameAndItemNumber(
