+++ tests/cypress/support/step_definitions/jupyterlite/I see JupyterLite Transformation dialog.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I see JupyterLite Transformation dialog", () => {
     new MaterialDesignerPage().designerWidget.jupyterLiteTransformationDialog.waitForVisible();
