+++ tests/cypress/support/step_definitions/jupyterlite/I see code in the cell is.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 function normalizeText(code: string) {
     return code.replace(/\s/g, "");
