+++ tests/cypress/support/step_definitions/jupyterlite/I see file opened.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I see file {string} opened", (filename: string) => {
     const { jupyterLiteSession } = new MaterialDesignerPage().designerWidget;
