+++ tests/cypress/support/step_definitions/jupyterlite/I see kernel status is.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I see kernel status is {string}", (expectedStatus: string) => {
     const { jupyterLiteSession } = new MaterialDesignerPage().designerWidget;
