+++ tests/cypress/support/step_definitions/jupyterlite/I set code in the cell to.ts
@@ -1,6 +1,6 @@
 import { When } from "@badeball/cypress-cucumber-preprocessor";
 
-import MaterialDesignerPage from "../widgets/MaterialDesignerPage";
+import MaterialDesignerPage from "../../widgets/MaterialDesignerPage";
 
 When("I set code in the cell {string} to:", (index: string, code: string) => {
     new MaterialDesignerPage().designerWidget.jupyterLiteSession.setCodeInCell(
